sval: Use uintptr_t for SVal internal representation
That way, the size of an SVal can definitely accommodate a
pointer, and will not be larger than the word size in 32-bit
architectures.

+++ src/Memory.hpp
@@ -52,10 +52,7 @@ namespace llvm {
 class SAddr {
 
 public:
-	/* The implementation is currently not quite portable.
-	 * Ideally, this should be equal to the size of a pointer
-	 * so that we can easily convert to actual pointers */
-	using Width = uint64_t;
+	using Width = uintptr_t;
 
 protected:
 	friend class llvm::Interpreter;
